requirements.txt
openai
python-dotenv
numpy
scipy
PyWavelets
google-generativeai
scikit-learn
PyYAML
dotenv
pandas
h5py
PyEMD
vmdpy
lancedb
umap-learn
networkx
matplotlib
PyPDF2
smolagents[all]
google-adk litellm
langchain
langchain_chroma
langchain_huggingface